pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tiktok-product-scout"
version = "1.0.0"
description = "Agentic automation tool for discovering trending products on TikTok before market saturation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "TikTok Product Scout Team"}
]
keywords = ["tiktok", "ecommerce", "dropshipping", "product-research", "automation", "web-scraping"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web scraping
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",

    # Data processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "rapidfuzz>=3.5.0",

    # API framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Google Trends
    "pytrends>=4.9.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "aiofiles>=23.2.0",

    # Logging and monitoring
    "loguru>=0.7.0",
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "ipython>=8.17.0",
]

redis = [
    "redis>=5.0.0",
    "celery>=5.3.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/tiktok-product-scout"
Repository = "https://github.com/yourusername/tiktok-product-scout"
Issues = "https://github.com/yourusername/tiktok-product-scout/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"
